+++ palace/drivers/eigensolver.cpp
@@ -148,13 +148,13 @@ EigenSolver::Solve(const std::vector<std::unique_ptr<Mesh>> &mesh) const
   std::unique_ptr<Operator> KM;
   if (iodata.solver.eigenmode.mass_orthog)
   {
-    // Mpi::Print(" Basis uses M-inner product\n");
-    // KM = spaceop.GetInnerProductMatrix(0.0, 1.0, nullptr, M.get());
-    // eigen->SetBMat(*KM);
-
-    Mpi::Print(" Basis uses (K + M)-inner product\n");
-    KM = spaceop.GetInnerProductMatrix(1.0, 1.0, K.get(), M.get());
+    Mpi::Print(" Basis uses M-inner product\n");
+    KM = spaceop.GetInnerProductMatrix(0.0, 1.0, nullptr, M.get());
     eigen->SetBMat(*KM);
+
+    // Mpi::Print(" Basis uses (K + M)-inner product\n");
+    // KM = spaceop.GetInnerProductMatrix(1.0, 1.0, K.get(), M.get());
+    // eigen->SetBMat(*KM);
   }
 
   // Construct a divergence-free projector so the eigenvalue solve is performed in the space
@@ -272,6 +272,14 @@ EigenSolver::Solve(const std::vector<std::unique_ptr<Mesh>> &mesh) const
       spaceop.GetMaterialOp(), spaceop.GetNDSpace(), iodata.solver.linear.estimator_tol,
       iodata.solver.linear.estimator_max_it, 0);
   ErrorIndicator indicator;
+  if (!KM)
+  {
+    // Normalize the finalized eigenvectors with respect to mass matrix (unit electric field
+    // energy) even if they are not computed to be orthogonal with respect to it.
+    KM = spaceop.GetInnerProductMatrix(0.0, 1.0, nullptr, M.get());
+    eigen->SetBMat(*KM);
+    eigen->RescaleEigenvectors(num_conv);
+  }
   for (int i = 0; i < iodata.solver.eigenmode.n; i++)
   {
     eigen->GetEigenvector(i, E);