+++ palace/linalg/arpack.hpp
@@ -66,8 +66,8 @@ protected:
   // Storage for Arnoldi basis vectors.
   std::unique_ptr<std::complex<double>[]> V;
 
-  // Storage for computed residual norms.
-  std::unique_ptr<double[]> res;
+  // Storage for computed residual norms and eigenvector scalings.
+  std::unique_ptr<double[]> res, xscale;
 
   // On input used to define optional initial guess, on output stores final residual
   // vector.
@@ -86,6 +86,9 @@ protected:
   // which case identity is used.
   const Operator *opB;
 
+  // Workspace vector for operator applications.
+  mutable ComplexVector x1, y1, z1;
+
   // Perform the ARPACK RCI loop.
   int SolveInternal(int n, std::complex<double> *r, std::complex<double> *V,
                     std::complex<double> *eig, int *perm);
@@ -96,6 +99,13 @@ protected:
   // Helper routines for ARPACK RCI.
   virtual void ApplyOp(const std::complex<double> *px, std::complex<double> *py) const = 0;
   virtual void ApplyOpB(const std::complex<double> *px, std::complex<double> *py) const = 0;
+
+  // Helper routine for computing the eigenvector normalization.
+  double GetEigenvectorNorm(const ComplexVector &x, ComplexVector &Bx) const;
+
+  // Helper routine for computing the eigenpair residual.
+  virtual double GetResidualNorm(std::complex<double> l, const ComplexVector &x,
+                                 ComplexVector &r) const = 0;
 
   // Helper routine for computing the backward error.
   virtual double GetBackwardScaling(std::complex<double> l) const = 0;
@@ -155,11 +165,17 @@ public:
   // Get the corresponding eigenvalue.
   std::complex<double> GetEigenvalue(int i) const override;
 
-  // Get the corresponding eigenvector.
+  // Get the corresponding eigenvector. Eigenvectors are normalized such that ||x||₂ = 1,
+  // unless the B-matrix is set for weighted inner products.
   void GetEigenvector(int i, ComplexVector &x) const override;
 
   // Get the corresponding eigenpair error.
   double GetError(int i, ErrorType type) const override;
+
+  // Re-normalize the given number of eigenvectors, for example if the matrix B for weighted
+  // inner products has changed. This does not perform re-orthogonalization with respect to
+  // the new matrix, only normalization.
+  void RescaleEigenvectors(int num_eig) override;
 };
 
 // Generalized eigenvalue problem solver: K x = λ M x .
@@ -172,12 +188,12 @@ private:
   // Operator norms for scaling.
   mutable double normK, normM;
 
-  // Workspace vector for operator applications.
-  mutable ComplexVector x, y, z;
-
 protected:
   void ApplyOp(const std::complex<double> *px, std::complex<double> *py) const override;
   void ApplyOpB(const std::complex<double> *px, std::complex<double> *py) const override;
+
+  double GetResidualNorm(std::complex<double> l, const ComplexVector &x,
+                         ComplexVector &r) const override;
 
   double GetBackwardScaling(std::complex<double> l) const override;
 
@@ -204,15 +220,14 @@ private:
   mutable double normK, normC, normM;
 
   // Workspace vectors for operator applications.
-  mutable ComplexVector x1, x2, y1, y2, z;
-
-  // Do eigenvector extraction from the linearized problem to the actual eigenvectors.
-  void ExtractEigenvector(std::complex<double> l, const std::complex<double> *py,
-                          std::complex<double> *px) const;
+  mutable ComplexVector x2, y2;
 
 protected:
   void ApplyOp(const std::complex<double> *px, std::complex<double> *py) const override;
   void ApplyOpB(const std::complex<double> *px, std::complex<double> *py) const override;
+
+  double GetResidualNorm(std::complex<double> l, const ComplexVector &x,
+                         ComplexVector &r) const override;
 
   double GetBackwardScaling(std::complex<double> l) const override;